+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.6.1] - 2025. 03. 19
 
 * Fix scrollbar drag performance issue
+* Add export service to export grid data to csv, pdf, json
 
 ## [1.6.0] - 2025. 03. 18
 
@@ -10,7 +11,7 @@
 * Add filter to the demo screen to make it easier to find a feature demo code.
 * Add some gif images to the documentation, document more features
 * add onLazyFetchCompleted event
-* add initial export service to export grid data to csv, excel, pdf, json ( under development )
+* add initial export service to export grid data to csv, pdf, json ( under development )
 * Update the documentation for boolean columns
 * Add scrollbar thumb hover color